Convert vector tile endpoint to ninja (#225)
