+++ django/src/rdwatch/views/__init__.py
@@ -1,4 +1,13 @@
 from .model_run import ModelRunViewSet
+from .satellite_image import (
+    all_satellite_timestamps,
+    satelliteimage_raster_bbox,
+    satelliteimage_raster_tile,
+    satelliteimage_time_list,
+    satelliteimage_visual_bbox,
+    satelliteimage_visual_tile,
+    satelliteimage_visual_time_list,
+)
 from .server_status import RetrieveServerStatus
 from .site_evaluation import site_evaluations
 from .site_model import post_region_model, post_site_model
@@ -7,16 +16,7 @@ from .site_observation import (
     get_site_observation_images,
     site_observations,
 )
-from .tile import (
-    all_satellite_timestamps,
-    satelliteimage_raster_bbox,
-    satelliteimage_raster_tile,
-    satelliteimage_time_list,
-    satelliteimage_visual_bbox,
-    satelliteimage_visual_tile,
-    satelliteimage_visual_time_list,
-    vector_tile,
-)
+from .vector_tile import vector_tile
 
 __all__ = [
     'ModelRunViewSet',
